+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.6.3
 - Fix standard library ignoring ignores 3rd-party libraries under site-packages/
+- Allow pathlib.Path object to be used in ignore items
 
 ## v0.6.2
 - Remove argument `full` for `nameof`, use `vars_only` instead. When `vars_only=False`, source of the argument returned.